+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The complete changelog for the Costs to Expect REST API, our changelog follows the format defined at https://keepachangelog.com/en/1.0.0/
 
+## [0.17.0] - [2022-08-01]
+### Added
+- Added a modal that displays your final score.
+### Changed
+- Adjusted the timer for the player scores, down to ten seconds from thirty.
+
 ## [0.16.0] - [2022-08-01]
 ### Added
 - Added ability to remove a player from an open game, deletes the score sheet, share token and assignment.